Library Reference

Tonti Diagrams

Decapods.TontiDiagrams.TheoryTontiDiagram — Constant

ACSet definition for a Tonti diagram.

This diagram is visualized in the q.uiver framework here. This Tonti diagram definition is currently very close to the simulation level and generalized away from the DEC tooling. Future work will be done to provide a closer integration between Tonti diagrams and DEC.

See Catlab.jl documentation for a description of the @present syntax.

Decapods.TontiDiagrams.Space — Type

Space

Structure for storing the computed values for the primal/dual complex, the boundary operators, the hodge stars, the laplacian operators, and the lie operators. This is the operator which provides the DEC connection for the Tonti diagra tooling.

Accessing the boundary, hodge, and laplacian operators is slightly complicated by using arrays in a 1-indexed system. The hodge and boundary operators are accessed with two indices, where the first determines the complex (1 for primal and 2 for dual) and the second determines the dimension (1 for 0-forms, 2 for 1-forms, 3 for 2-forms, etc.). Thus, the boundary operator from dual 1-forms to dual 2-forms is given as:

sp = Space(s)
sp.boundary[2,2]

The other goal of this structure is to cache the computed operators, yet optimally this caching should be perfomed during the vectorfield operation.

Decapods.TontiDiagrams.Space — Method

Space(s::EmbeddedDeltaSet3D)

Caclulates all of the values stored in the Space object for a given complex s.

Decapods.TontiDiagrams.TontiDiagram — Method

TontiDiagram()

Initialize an empty TontiDiagram object.

Decapods.TontiDiagrams.Open — Method

Open(td::TontiDiagram, states::Symbol)

Generates an OpenTontiDiagram with cospan legs on variables defined by the symbols included in states. This OpenTontiDiagram can then be composed with other OpenTontiDiagrams over a pattern given by an undirected wiring diagram.

OpenTontiDiagram(td, :x, :v)

Decapods.TontiDiagrams.add_bc! — Method

addbc!(td::TontiDiagram, varsys::Symbol, func::Function)

Adds a "boundary condition" to the variable var_sys by applying func to the values of this variable during simulation. This function is the last one evaluated on the data of the variable var_sys, and so can be used to enforce any relevant boundary conditions.

TODO: Add time dependency of boundary condition function to allow for time-varying BCs.

Decapods.TontiDiagrams.add_derivative! — Method

addderivative!(td::TontiDiagram, sp::Space, domsym::Symbol, codom_sym::Symbol)

Adds a derivative transition from variable dom_sym to variable codom_sym using the boundary operators from sp. This function determines which boundary operator to use and inserts an appropriate transition between the two variables.

Defining a spatial derivative relationship between the primal 0-form x and the primal 1-form Δx cis given as follows:

add_derivative(td, sp, :x, :Δx)

Decapods.TontiDiagrams.add_derivatives! — Method

add_derivatives!(td::TontiDiagram, sp::Space, vars:Pair{Symbol, Symbol}...)

Adds multiple derivative transition between pairs of variables, using the same syntax as in add_derivative!.

Example usage:

add_derivatives!(td, sp, (:x,:Δx), (:y, :Δy))

Decapods.TontiDiagrams.add_laplacian! — Method

addlaplacian!(td::TontiDiagrams, sp::Space, domsym::Symbol, codom_sym::Symbol; coef::Float64)

Adds a transition which defines codom_sym as the laplacian of dom_sym with a constant scaling factor of coef.

Decapods.TontiDiagrams.add_time_dep! — Method

addtimedep!(td::TontiDiagram, derivsym::Symbol, integsym::Symbol)

Adds a time derivative relationship between the variables deriv_sym and integ_sym (where deriv_sym is the time derivative of integ_sym). These relationships are used to determine the state-variables of the system.

Decapods.TontiDiagrams.add_transition! — Method

addtransition!(td, domsym::Vector{Symbol}, func!, codom_sym::Vector{Symbol})

Adds a transition function from variables dom_sym to variables codom_sym with its transition defined by func!. func! is expected to have the signature func(codom_sym..., dom_sym...) and is expected to modify the values of the codom_sym variables.

Defining a transition from variables x and y to m that calculates the magnitude of the values in x and y as vector coordinates would be defined as:

td = TontiDiagram()
add_variables!(td, (:x, 0, 1), (:y, 0, 1). (:m, 0, 1))
add_transition!(td, [:x, :y], (m,x,y)->(m .= sqrt.(x .* y)), [:m])

Decapods.TontiDiagrams.add_variable! — Method

add_variable!(td:TontiDiagram, symbol::Symbol, dimension::Int64, complex::Bool)

Adds a variable to the TontiDiagram system which can later be referenced by its symbol. This constructor requires the dimensionality of the variable (0 -> point, 1 -> line, etc.) and the complex it is defined on (true -> primal/straight, false -> dual/twisted).

Definiting a system with a variable v defined on the primal lines would be constructed as:

td = TontiDiagram()
add_variable!(td, :v, 1, true)

Decapods.TontiDiagrams.add_variables! — Method

add_variables!(td:TontiDiagram, vars::Tuple{Symbol, Int64, Bool}...)

Adds multiple variables to the TontiDiagram system which can later be referenced by their symbols. This constructor follows the same pattern as add_variable! with each variable specified as a tuple of:

(symbol, dimension, complex)

Defining a system with a variable v defined on the primal lines, a variable p̃ defined on the dual surfaces, and a variable C defined on the primal surfaces would be constructed as:

td = TontiDiagram()
add_variables!(td, (:v, 1, true), (:p̃, 1, true), (:C, 2, true))

Decapods.TontiDiagrams.gen_form — Method

Construct a 0-form based on a scalar function

This operator accepts a scalar function and evaulates it at each point on the simplex, returning a 0-form.

Decapods.TontiDiagrams.vectorfield — Method

vectorfield(td::AbstractTontiDiagram, sp::Space)

Generates a Julia function which calculates the vectorfield of the Tonti diagram. The state of the system is defined by a single vector which is a flattening of all state variables of the system. Thus, this function returns both the indices of each variable in the state-vector along with the vectorfield function itself.

The resulting function has a signature of the form f!(du, u, p, t) and can be passed to the DifferentialEquations.jl solver package.