Library Reference

Diagrams

Catlab.CategoricalAlgebra.Diagrams — Module

Diagrams in a category and their morphisms.

Catlab.CategoricalAlgebra.Diagrams.Diagram — Type

Diagram in a category.

Recall that a diagram in a category $C$ is a functor $D: J → C$, where for us the shape category $J$ is finitely presented. Although such a diagram is captured perfectly well by a FinDomFunctor, there are several different notions of morphism between diagrams. The first type parameter T in this wrapper type distinguishes which diagram category the diagram belongs to. See DiagramHom for more about the possible choices. The parameter T may also be Any to indicate that no choice has (yet) been made.

Catlab.CategoricalAlgebra.Diagrams.DiagramHom — Type

Morphism of diagrams in a category.

In fact, this type encompasses several different kinds of morphisms from a diagram $D: J → C$ to another diagram $D′: J′ → C$:

DiagramHom{id}: a functor $F: J → J′$ together with a natural transformation $ϕ: D ⇒ F⋅D′$

DiagramHom{op}: a functor $F: J′ → J$ together with a natural transformation $ϕ: F⋅D ⇒ D′$

DiagramHom{co}: a functor $F: J → J′$ together with a natural transformation $ϕ: F⋅D′ ⇒ D$.

Note that Diagram{op} is not the opposite category of Diagram{id}, but Diagram{op} and Diagram{co} are opposites of each other. Explicit support is included for both because they are useful for different purposes: morphisms of type DiagramHom{id} and DiagramHom{op} induce morphisms between colimits and between limits of the diagrams, respectively, whereas morphisms of type DiagramHom{co} generalize morphisms of polynomial functors.

Catlab.CategoricalAlgebra.Diagrams.QueryDiagram — Type

Diagram representing a (conjunctive or gluing) query.

Besides the diagram functor itself, a query diagram contains a dictionary of query parameters.

Catlab.CategoricalAlgebra.Diagrams.SimpleDiagram — Type

Default concrete type for diagram in a category.

Catlab.CategoricalAlgebra.Diagrams.diagram — Method

Functor underlying a diagram object.

Catlab.CategoricalAlgebra.Diagrams.shape — Method

The shape or indexing category of a diagram.

This is the domain of the underlying functor.

Open Diagrams

Modules = [ Decapodes.OpenDiagrams ]
Private = false

Simulation

Modules = [ Decapodes.Simulations ]
Private = false

Schedules

AlgebraicRewriting.Schedules.apply_schedule — Method

apply schedule and return whether or not the input changed

Examples

Modules = [ Decapodes.Examples ]
Private = false

PetriNets

Modules = [ Decapodes.PetriNets]
Private = false