Library Reference

Decapodes

Decapodes.compile — Method

compile(d::SummationDecapode, inputs::Vector{Symbol}, alloc_vectors::Vector{AllocVecCall}, optimizable_dec_operators::Set{Symbol}, dimension::Int, stateeltype::DataType, code_target::GenerationTarget)

Function that compiles the computation body. d is the input Decapode, inputs is a vector of state variables and literals, alloc_vec should be empty when passed in, optimizable_dec_operators is a collection of all DEC operator symbols that can use special in-place methods, dimension is the dimension of the problem (usually 1 or 2), stateeltype is the type of the state elements (usually Float32 or Float64) and code_target determines what architecture the code is compiled for (either CPU or CUDA).

Decapodes.compile_env — Method

compile_env(d::SummationDecapode, dec_matrices::Vector{Symbol}, con_dec_operators::Set{Symbol}, code_target::GenerationTarget)

This creates the symbol to function linking for the simulation output. Those run through the default_dec backend expect both an in-place and an out-of-place variant in that order. User defined operations only support out-of-place.

Decapodes.gensim — Method

gensim(user_d::SummationDecapode, input_vars::Vector{Symbol}; dimension::Int=2, stateeltype::DataType = Float64, code_target::GenerationTarget = CPUTarget())

Generates the entire code body for the simulation function. user_d is the user passed Decapodes which will be left unmodified and 'input_vars' is the collection of state variables and literals in the Decapode.

Optional keyword arguments are dimension, which is the dimension of the problem and defaults to 2D, stateeltype, which is the element type of the state forms and defaults to Float64 and code_target, which is the intended architecture target for the generated code, defaulting to regular CPU compatible code.

Decapodes.gensim — Method

function gensim(d::SummationDecapode; dimension::Int=2)

Generate a simulation function from the given Decapode. The returned function can then be combined with a mesh and a function describing function mappings to return a simulator to be passed to solve.

[ Info: Page built in 2 seconds.
[ Info: This page was last built at 2024-06-13T19:54:46.887.