Library Reference

Decapodes

Decapodes.op1_inf_rules_1D — Constant

These are the default rules used to do type inference in the 1D exterior calculus.

Decapodes.op1_inf_rules_2D — Constant

These are the default rules used to do type inference in the 2D exterior calculus.

Decapodes.op1_res_rules_1D — Constant

These are the default rules used to do function resolution in the 1D exterior calculus.

Decapodes.op1_res_rules_2D — Constant

These are the default rules used to do function resolution in the 2D exterior calculus.

Decapodes.CartesianPoint — Type

CartesianPoint{T}(p)

a point in cartesian coordinates, intended as a wrapper around Point3 from GeometryBasics.

Decapodes.SpherePoint — Type

SpherePoint{T}(p)

a point in spherical coordinates, intended as a wrapper around Point3 from GeometryBasics.

Decapodes.TangentBasis — Method

tb(w)

Take a linear combinations of the tangent vectors at the base point. Use this to get a vector tangent to the sphere in the coordinate system of the base point. If the base point is in spherical coordinates, this is the identity, if the base point is in cartesian coordinates, it returns the tangent vector in cartesian coordinates.

Catlab.WiringDiagrams.WiringDiagramAlgebras.oapply — Method

function oapply(relation::RelationDiagram, podes::Vector{D}) where {D<:OpenSummationDecapode}

Compose a list of decapodes as specified by the given relation diagram.

The decapodes must be given in the same order as they were specified in the relation.

State variables (such as the (C,V) given in the head of the following @relation) do not affect the result of a composition.

Examples

julia> compose_diff_adv = @relation (C,V) begin
  diffusion(C, ϕ₁)
  advection(C, ϕ₂, V)
  superposition(ϕ₁, ϕ₂, ϕ, C)
end;

julia> oapply(compose_diff_adv, [(Diffusion, [:C, :ϕ]),
  (Advection, [:C, :ϕ, :V]), (Superposition, [:ϕ₁, :ϕ₂, :ϕ, :C])]);

Decapodes.Open — Method

Open(d::SummationDecapode{T,U,V}, names::AbstractVector{Symbol}) where {T,U,V}

creates an OpenSummationDecapode based on named variables rather than variable indices. See AlgebraicPetri.jl's Open for the analogous verion for LabelledReactionNetworks.

Decapodes.contract_operators — Method

function contract_operators(d::SummationDecapode)

Find chains of Op1s in the given Decapode, and replace them with a single Op1 with a vector of function names. After this process, all Vars that are not a part of any computation are removed.

Decapodes.dot_rename! — Method

dot_rename!(d::AbstractNamedDecapode)

Rename tangent variables by their depending variable appended with a dot. e.g. If D == ∂ₜ(C), then rename D to Ċ.

If a tangent variable updates multiple vars, choose one arbitrarily. e.g. If D == ∂ₜ(C) and D == ∂ₜ(B), then rename D to either Ċ or B ̇.

Decapodes.expand_operators — Method

function expand_operators(d::SummationDecapode)

Find operations that are compositions, and expand them with intermediate variables.

Decapodes.fill_names! — Method

fill_names!(d::AbstractNamedDecapode)

Provide a variable name to all the variables that don't have names.

Decapodes.gensim — Method

function gensim(d::AbstractNamedDecapode; dimension::Int=2)

Generate a simulation function from the given Decapode. The returned function can then be combined with a mesh and a function describing function mappings to return a simulator to be passed to solve.

Decapodes.infer_types! — Method

function infer_types!(d::SummationDecapode, op1_rules::Vector{NamedTuple{(:src_type, :tgt_type, :replacement_type, :op), NTuple{4, Symbol}}})

Infer types of Vars given rules wherein one type is known and the other not.

Decapodes.loadmesh — Function

loadmesh(s, subdivision=Circumcenter())

Load in a mesh from the Artifacts.jl system and automatically subdivide it.

Decapodes.loadmesh — Method

loadmesh(s::Icosphere)

Load in a icosphere mesh.

Decapodes.loadmesh — Method

loadmesh(s::Point_Map)

Load in a point map describing the connectivity of the toroidal mesh.

Decapodes.loadmesh — Method

loadmesh(s::Rectangle_30x10)

Load in a rectangular mesh.

Decapodes.loadmesh — Method

loadmesh(s::Torus_30x10)

Load in a toroidal mesh.

Decapodes.recursive_delete_parents — Method

function recursive_delete_parents!(d::SummationDecapode, to_delete::Vector{Int64})

Delete the given nodes and their parents in the decapode, recursively.

Decapodes.resolve_overloads! — Method

function resolve_overloads!(d::SummationDecapode, op1_rules::Vector{NamedTuple{(:src_type, :tgt_type, :resolved_name, :op), NTuple{4, Symbol}}})

Resolve function overloads based on types of src and tgt.

Decapodes.resolve_overloads! — Method

function resolve_overloads!(d::SummationDecapode)

Resolve function overloads based on types of src and tgt.

Decapodes.unicode! — Method

function unicode!(d::SummationDecapode)

Replace ASCII operators with their Unicode equivalents.

Decapodes.unique_by! — Method

function unique_by!(acset, column_names::Vector{Symbol})

Given column names from the same table, remove duplicate rows.

WARNING: This function does not check if other tables index into the one given. Removal of rows is performed with prejudice.

See also: unique_by.

Examples

julia> unique_by!(parallel_arrows(Graph, 123), :E, [:src,:tgt]) == parallel_arrows(Graph, 1)
true

Decapodes.unique_by — Method

function unique_by(acset, column_names::Vector{Symbol})

Given column names from the same table, return a copy of the acset with duplicate rows removed. Removal of rows is performed with prejudice.

WARNING: This function does not check if other tables index into the one given. Removal of rows is performed with prejudice.

See also: unique_by!.

Examples

julia> unique_by(parallel_arrows(Graph, 123), :E, [:src,:tgt]) == parallel_arrows(Graph, 1)
true

Decapodes.vec_to_dec! — Method

function vec_to_dec!(d::SummationDecapode)

Replace Vector Calculus operators with Discrete Exterior Calculus equivalents.

Decapodes.@decapode — Macro

macro decapode(e)

Construct a Decapode.