Library Reference

Diagrams

Decapodes.Diagrams — Module

Provides the tooling for defining Decapodes conveniently with symbolic expressions

Decapodes.Diagrams.Decapodes2D — Constant

Decapodes2D A schema which includes any homomorphisms that may be added by the @decapode macro.

TODO: This should be chipped away at as more of this tooling takes advantage of the Catlab GAT system

Decapodes.Diagrams.@decapode — Macro

decapode(cat, body)

This macro is called as:

@decapode presentation begin
  expressions
end

where presentation is a Presentation containing the syntax for the Decapode (operators and form types), and expressions is symbolic expressions of the relations within the Decapode.

Open Diagrams

Decapodes.OpenDiagrams.OpenDiagram — Type

Open diagram as a structured multicospan in R-form.

An open diagram is a diagram, represented as a FinDomFunctor, together with a cospan of finite sets whose apex is the object set of the diagram's indexing category.

Simulation

Decapodes.Simulations.ArbitraryFunc — Type

ArbitraryFunc

Out of place function which applies to any number of arguments.

Decapodes.Simulations.ConstantFunc — Type

ConstantFunc

Function which always returns the same value.

Decapodes.Simulations.ElementwiseFunc — Type

ElementwiseFunc

Function applied to each element of one or more vectors.

Decapodes.Simulations.InPlaceFunc — Type

InPlaceFunc

Function which operates on arguments in-place, not returning a value.

Decapodes.Simulations.MatrixFunc — Type

MatrixFunc

Linear operator applied through matrix multiplication.

Decapodes.Simulations.TDInPlaceFunc — Type

TDInPlaceFunc

Function which accepts the current time as the last argument. Used to set up time-dependent physics or boundary conditions.

Decapodes.Simulations.gen_sim — Method

gen_sim(dwd::WiringDiagram, name2func::Dict, s::EmbeddedDeltaDualComplex2D; form2dim=form2dim, params=[], autodiff=false)

This function generates a function which evaluates the acylcic directed wiring diagram dwd, using the functions in name2func for operators to execute corresonding to each box and information from the mesh s to pre-allocate necessary memory. This operator can generate a function which is compatible with the autodifferentiation solvers in DifferentialEquations.jl.

Schedules

Decapodes.Schedules — Module

Schedules

This module contains the tooling which converts diagrams to DWDs.

TODO This module will also include optimization tooling relating to DWD schedules (compressing matrix operations, pre-computing constant values, etc).

Decapodes.Schedules.diag2dwd — Method

diag2dwd(diagram; clean = false, calcstates = [], outvars=[], in_vars=[])

Generates a directed wiring diagram (DWD) from a Decapode diagram. This method generates a DWD which represents a single explicit time step of the system, computed by the dependency graph of each time derivative within the system. This does not necessarily work for all Decapodes, and only generates an explicit time-step solution.

TODO: Functions like this can be created to generate other solution styles, like implicit solutions.

Examples

Decapodes.Examples.dual — Method

dual(s::EmbeddedDeltaSet2D{O, P})

Generates a dual mesh for the provided delta set s.

TODO: This tooling should be upstreamed to CombinatorialSpaces

Decapodes.Examples.expand_dwd — Method

expanddwd(dwdorig::WiringDiagram, patterns::Dict{Symbol, WiringDiagram})

Replaces any boxes in dwd_orig which map to a key in patterns with their corresponding diagram in patterns. This allows for replacement of complex operations with their definition.

PetriNets

Decapodes.PetriNets.expand_pres! — Method

expand_pres!(pres::Presentation, pn::LabelledReactionNet)

Expands the syntax of the presentation pres to include the necessary operators for expressing mass action on the Petri net pn.

Decapodes.PetriNets.gen_functions — Method

gen_functions(pn::LabelledReactionNet, s::EmbeddedDeltaDualComplex2D)

Generates the functions necessary for executing the Petri net after it has been converted to a Decapode.

Decapodes.PetriNets.pn2dec — Method

pn2dec(prs::Presentation, pn::LabelledReactionNet)

Generates a Decapode diagram which represents the law of mass action applied to the petri net pn with the syntax from the presentation prs.